Support GRU go_backwards=True
